requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.1.5
attrs==19.1.0
Automat==0.7.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
blinker==1.4
bzr==2.8.0.dev1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Cheetah3==3.1.0
click==7.0.dev0
colorama==0.4.1
configobj==5.0.6
configparser==3.5.0b2
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cvxopt==1.1.9
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20121119
distro==1.3.0
entrypoints==0.2.3.post3
enum34==1.1.6
et-xmlfile==1.0.1
fasteners==0.14.1
Flask==1.0.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
fuzzywuzzy==0.17.0
gyp==0.1
html5lib==1.0.1
httplib2==0.11.3
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
intelhex==2.2.1
intervaltree==2.1.0
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
jdcal==1.0
Jinja2==2.10
joblib==0.12.2
jsonschema==2.6.0
junit-xml==1.8
keyring==15.1.0
keyrings.alt==3.1
kiwisolver==1.0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lockfile==0.12.2
logilab-common==1.4.1
lxml==4.3.2
lz4==1.1.0
MarkupSafe==1.1.0
matplotlib==2.2.2
mbed-cli==1.8.3
mbed-greentea==1.5.0
mbed-host-tests==1.4.4
mbed-ls==1.6.2
mccabe==0.6.1
mercurial==4.6.1
mock==2.0.0
monotonic==1.5
more-itertools==4.2.0
networkx==2.1
numexpr==2.6.5
numpy==1.14.5
oauthlib==2.0.6
olefile==0.46
openpyxl==2.4.9
pandas==0.23.3
parsel==1.5.1
patsy==0.4.1+dev
pbr==5.1.1
Pillow==5.4.1
pluggy==0.6.0
prettytable==0.7.2
py==1.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
pyelftools==0.25
Pygments==2.2.0
PyGObject==3.30.1
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.6.4
pyliblzma==0.5.3
pylint==1.9.2
pyocd==0.14.3
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyserial==3.4
pytest==3.6.4
python-dateutil==2.6.1
python-pptx==0.6.17
pytz==2018.5
pyusb==1.0.2
pyxdg==0.25
PyYAML==3.13
pyzmq==17.1.0
queuelib==1.5.0
requests==2.21.0
scipy==1.1.0
Scrapy==1.6.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.15.0
singledispatch==3.4.0.3
six==1.12.0
SOAPpy==0.12.22
sortedcontainers==2.1.0
statsmodels==0.8.0
subprocess32==3.5.2
tables==3.4.4
Twisted==18.9.0
urllib3==1.24.1
vcstools==0.1.40
w3lib==1.20.0
wadllib==1.3.3
webencodings==0.5
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.9.0
wstool==0.1.17
wstools==0.4.3
WTForms==2.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xlrd==1.1.0
XlsxWriter==1.1.5
xlwt==1.3.0
zope.interface==4.6.0
